Change GroupedAccumulatorState groupId from long to int

+++ core/trino-main/src/main/java/io/trino/operator/aggregation/DistinctAccumulatorFactory.java
@@ -202,7 +202,7 @@ public class DistinctAccumulatorFactory
         }
 
         @Override
-        public void setGroupCount(long groupCount)
+        public void setGroupCount(int groupCount)
         {
             accumulator.setGroupCount(groupCount);
         }
