+++ core/trino-main/src/main/java/io/trino/operator/aggregation/state/TDigestStateFactory.java
@@ -44,7 +44,7 @@ public class TDigestStateFactory
         private long size;
 
         @Override
-        public void ensureCapacity(long size)
+        public void ensureCapacity(int size)
         {
             digests.ensureCapacity(size);
         }
